Move non-UI files from `ui/` to `model`
