+++ app/src/androidTest/java/com/github/se/signify/ui/screens/BottomBarNavigationTest.kt
@@ -3,9 +3,9 @@ package com.github.se.signify.ui.screens
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
+import com.github.se.signify.model.navigation.LIST_TOP_LEVEL_DESTINATION
+import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.ui.navigation.BottomNavigationMenu
-import com.github.se.signify.ui.navigation.LIST_TOP_LEVEL_DESTINATION
-import com.github.se.signify.ui.navigation.NavigationActions
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
