+++ app/src/androidTest/java/com/github/se/signify/ui/screens/UtilsTest.kt
@@ -16,6 +16,7 @@ import androidx.test.rule.GrantPermissionRule
 import com.github.se.signify.R
 import com.github.se.signify.model.di.AppDependencyProvider
 import com.github.se.signify.model.hand.HandLandMarkViewModel
+import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.ui.AccountInformation
 import com.github.se.signify.ui.AnnexScreenScaffold
 import com.github.se.signify.ui.BackButton
@@ -37,7 +38,6 @@ import com.github.se.signify.ui.UtilTextButton
 import com.github.se.signify.ui.getIconResId
 import com.github.se.signify.ui.getImageResId
 import com.github.se.signify.ui.getTipResId
-import com.github.se.signify.ui.navigation.NavigationActions
 import junit.framework.TestCase.assertEquals
 import org.junit.Rule
 import org.junit.Test
