+++ app/src/androidTest/java/com/github/se/signify/ui/screens/auth/UnauthenticatedScreenTest.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
-import com.github.se.signify.ui.navigation.NavigationActions
-import com.github.se.signify.ui.navigation.Screen
+import com.github.se.signify.model.navigation.NavigationActions
+import com.github.se.signify.model.navigation.Screen
 import org.junit.Rule
 import org.junit.Test
 import org.mockito.Mockito.mock
