+++ app/src/androidTest/java/com/github/se/signify/ui/screens/challenge/CreateAChallengeScreenTest.kt
@@ -6,8 +6,8 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import com.github.se.signify.model.auth.MockUserSession
 import com.github.se.signify.model.auth.UserSession
 import com.github.se.signify.model.challenge.MockChallengeRepository
+import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.user.UserRepository
-import com.github.se.signify.ui.navigation.NavigationActions
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
