+++ app/src/androidTest/java/com/github/se/signify/ui/screens/home/ExerciseScreenTest.kt
@@ -11,7 +11,7 @@ import androidx.test.rule.GrantPermissionRule
 import com.github.se.signify.model.di.AppDependencyProvider
 import com.github.se.signify.model.exercise.ExerciseLevel
 import com.github.se.signify.model.hand.HandLandMarkViewModel
-import com.github.se.signify.ui.navigation.NavigationActions
+import com.github.se.signify.model.navigation.NavigationActions
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
