+++ app/src/androidTest/java/com/github/se/signify/ui/screens/home/HomeScreenTest.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import com.github.se.signify.model.exercise.ExerciseLevel
-import com.github.se.signify.ui.navigation.NavigationActions
-import com.github.se.signify.ui.navigation.Screen
+import com.github.se.signify.model.navigation.NavigationActions
+import com.github.se.signify.model.navigation.Screen
 import kotlinx.coroutines.CoroutineScope
 import org.junit.Before
 import org.junit.Rule
