+++ app/src/main/java/com/github/se/signify/MainActivity.kt
@@ -19,9 +19,9 @@ import com.github.se.signify.model.di.AppDependencyProvider
 import com.github.se.signify.model.di.DependencyProvider
 import com.github.se.signify.model.exercise.ExerciseLevel
 import com.github.se.signify.model.hand.HandLandMarkViewModel
-import com.github.se.signify.ui.navigation.NavigationActions
-import com.github.se.signify.ui.navigation.Route
-import com.github.se.signify.ui.navigation.Screen
+import com.github.se.signify.model.navigation.NavigationActions
+import com.github.se.signify.model.navigation.Route
+import com.github.se.signify.model.navigation.Screen
 import com.github.se.signify.ui.screens.WelcomeScreen
 import com.github.se.signify.ui.screens.auth.LoginScreen
 import com.github.se.signify.ui.screens.auth.UnauthenticatedScreen
@@ -62,8 +62,7 @@ fun SignifyAppPreview(
     dependencyProvider: DependencyProvider,
 ) {
   val navController = rememberNavController()
-  val navigationActions =
-      NavigationActions(context, navController, dependencyProvider.userSession())
+  val navigationActions = NavigationActions(navController, dependencyProvider.userSession())
   val handLandMarkImplementation = dependencyProvider.handLandMarkRepository()
   val handLandMarkViewModel: HandLandMarkViewModel =
       viewModel(factory = HandLandMarkViewModel.provideFactory(context, handLandMarkImplementation))
