+++ app/src/main/java/com/github/se/signify/ui/Utils.kt
@@ -79,9 +79,9 @@ import androidx.core.content.ContextCompat
 import coil.compose.AsyncImage
 import com.github.se.signify.R
 import com.github.se.signify.model.hand.HandLandMarkViewModel
+import com.github.se.signify.model.navigation.LIST_TOP_LEVEL_DESTINATION
+import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.ui.navigation.BottomNavigationMenu
-import com.github.se.signify.ui.navigation.LIST_TOP_LEVEL_DESTINATION
-import com.github.se.signify.ui.navigation.NavigationActions
 
 // Map to associate each letter with its corresponding drawable resource for ASL gestures
 val gestureImageMap =
