+++ app/src/main/java/com/github/se/signify/ui/navigation/BottomNavigationMenu.kt
@@ -16,6 +16,8 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.github.se.signify.model.navigation.Screen
+import com.github.se.signify.model.navigation.TopLevelDestination
 
 @Composable
 fun BottomNavigationMenu(
